refactor(users): extract member directory + profile endpoints to _members.py
_auth.py had grown to the 500-line limit. Move the two member-facing read
endpoints (GET /users/directory/, GET /users/{id}/profile/) into a dedicated
_members.py router, registered alongside the other user sub-routers in api.py.
Pure move — no route paths, behavior, or response shapes change.
