+++ backend/users/api.py
@@ -11,6 +11,7 @@ from users._helpers import (  # re-exported
 from users._magic_links import router as magic_links_router
 from users._magic_login import router as magic_login_router
 from users._management import router as management_router
+from users._members import router as members_router
 from users._roles import router as roles_router
 
 __all__ = [
@@ -23,6 +24,7 @@ __all__ = [
 router = Router()
 router.add_router("", roles_router)
 router.add_router("", auth_router)
+router.add_router("", members_router)
 router.add_router("", magic_login_router)
 router.add_router("", management_router)
 router.add_router("", magic_links_router)
